+++ lib/AppInfo/Application.php
@@ -8,11 +8,14 @@ declare(strict_types=1);
 namespace OCA\Shortcloud\AppInfo;
 
 use OCA\Files\Event\LoadSidebar;
+use OCA\Shortcloud\Listener\AppChangedListener;
 use OCA\Shortcloud\Listener\LoadSidebarListener;
 use OCA\Shortcloud\Listener\ShareCreatedListener;
 use OCA\Shortcloud\Listener\ShareDeletedListener;
 use OCA\Shortcloud\Listener\UserDeletedListener;
 use OCA\Shortcloud\SetupCheck\RewriteCheck;
+use OCP\App\Events\AppEnableEvent;
+use OCP\App\Events\AppUpdateEvent;
 use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
@@ -33,6 +36,8 @@ class Application extends App implements IBootstrap {
 		$context->registerEventListener(ShareDeletedEvent::class, ShareDeletedListener::class);
 		$context->registerEventListener(LoadSidebar::class, LoadSidebarListener::class);
 		$context->registerEventListener(UserDeletedEvent::class, UserDeletedListener::class);
+		$context->registerEventListener(AppUpdateEvent::class, AppChangedListener::class);
+		$context->registerEventListener(AppEnableEvent::class, AppChangedListener::class);
 		$context->registerSetupCheck(RewriteCheck::class);
 	}